refactor: 공통적으로 사용되는 유틸 및 상수를 common 디렉토리로 이동
- 상위 모듈인 공통 훅에서 하위 모듈인 booking 유틸/상수에 의존하는 구조였음
- 의존성 방향이 거꾸로 설정된 문제를 해결하고, 재사용성을 높이기 위해 common으로 분리
