+++ frontend/src/common/constants/apiEndpoints.ts
@@ -2,4 +2,8 @@ export const API_ENDPOINTS = {
   SPECIALTIES: '/categories',
   MENTORINGS: '/mentorings',
   RESERVATION: '/reservation',
+  VALIDATE_ID: 'validate-id',
+  SIGNUP: '/signup',
+  AUTH_CODE: '/auth-code',
+  AUTH_CODE_VERIFY: '/auth-code/verify',
 } as const;
